[CDF-24304] 🫣 Mask sensitive output in terminal (#1499)
# Description

Issue is caused by comparing CDF to local resource and printing the diff
to the terminal. The solution I propose in this PR is to introduce the
method `.sensitive_strings()` to our `ResourceLoaders` (standardized
interface for working with CDF resources). Then, for all loaders that
deals with credentials (Transformations, FunctionSchedule,
WorkflowTrigger, HostedExtactorsSource), implement this method.

## Changelog

- [x] Patch
- [ ] Minor
- [ ] Skip

## cdf

### Fixed

- When running `cdf deploy --verbose`, Cognite Toolkit now masks
credentials in the terminal output.

## templates

No changes.