[CDF-24233, CDF-24234, CDFBUG-1025] 🚶🏻Dump command wrong model issue (#1500)
# Description

Interactive selection of data models can lead to the selection of a
model that is different than you selected. Ended up doing a
reimplementation of the `_interactive_select` for the DataModelsFinder.
This was to make it more use friendly. See comments in code.

## Changelog

- [x] Patch
- [ ] Minor
- [ ] Skip

## cdf

### Fixed

- When running `cdf dump datamodel` in interfactive mode, the Toolkit
now always dumps the selected data model. Earlier versions, depending on
the choices the user made, Toolkit could end up not dumping the correct
model.

## templates

No changes.