[CDF-24288] 🪸Dump resource issue (#1502)
# Description

All loaders (standardized interface to work with CDF resources) have a
`dump_resource` method. In some of the loaders, this method assumes that
the local representation of the resource has all required fields. This
is wrong as, for example, the `cdf modules pull` only requires the
identifier fields to be present.

In the first commit, created a test that tries to dump Response/Read
format of a resource together with local dict with only the identifier
keys. This triggers all the implementations of the `dump_resource` were
this wrong assumption is made.

## Bug screenshot

![image](https://github.com/user-attachments/assets/51cf3910-025a-4fa3-957e-9031030aabd0)


## Changelog

- [x] Patch
- [ ] Minor
- [ ] Skip

## cdf

### Fixed

- When using `cdf modues pull`, if a local resource is missing required
parameters, the Toolkit no longer raises a `AttributeError` or
`KeyError`. Instead, the resource is fetched from CDF and written to the
local configuration YAML.

## templates

No changes.