[CDF-24080] 👨🏼‍🏭Pull group missing capabilities (#1504)
# Description

This issue was caused by having two loaders (standardized interface to
work with CDF resources) to work with Groups (one for all scoped and one
of resource scoped). When pulling the group (retrieving the Group from
CDF and writing it to the local configuration), the first loader would
write correctly to the local file. While the second pass, would cause
some of the capabilities disappear. Solution is to only use one of the
loaders for groups.

Found a logic mistake in the if conditions for skipping unsupported
loaders, so fixed that issue as well.

## Changelog

- [x] Patch
- [ ] Minor
- [ ] Skip

## cdf

### Fixed

- Running `cdf modules pull` no longer skips capabilities for Group
resources.
- Running `cdf modules pull` now correctly skips unsupported resources
such as `HostedExtractor` `source` and `destination`, as well as
Functions and Streamlit apps.

## templates

No changes.