[CDF-24358] 🥡 cdf dump datamodel --include-global flag (#1506)
# Description

Put behind an alpha flag, planned for the `0.5` release. 

## Changelog

- [x] Patch
- [ ] Minor
- [ ] Skip

## cdf

### Added

- [alpha feature] The alpha flag `dump-dm-global` changes the behavior
of cdf dump datamodel to skip global resources. The motivation for this
change is the following observed user flow. First, a data model is
developed outside Toolkit (neat/CDF UI). Then, this command is used to
create the configuration that are checked into version control history.
In this case, global resources should not be part of the configuration.
For example, a CogniteCore extension model should not include the
Cognite containers and views in the local configurations as these are
controlled by Cognite.

## templates

No changes.