[CDF-24392] 🚧 Simplify extraction pipeline config loader (#1517)
# Description

**Context** Loader is a standardized interface we use in Toolkit to work
with CDF resources. The `ExtractionPipelineConfig` was written very
early in Toolkit. Now, we have standardized the logic that compares the
CDF resources vs the local resource and do deployment if it has changed.
This PR simplifies the `.create`/`.update` methods for the
`ExtractionPipelineConfig` removing the compare logic which is no longer
necessary.


![image](https://github.com/user-attachments/assets/ddcbf7ba-25a6-4b84-ba8a-751280922a02)


## Changelog

- [x] Patch
- [ ] Minor
- [ ] Skip

## cdf

### Fixed

- When running `cdf deploy`, Toolkit no longer warns about missing
Environment variables inside the `config` field of Extraction Pipeline
Config resources.

## templates

No changes.