[CDF-24270] 🍂 cdf modules add withouth an config.yaml (#1515)
# Description

The `config.[env].yaml` file is optional, but was unfortunately
hardcoded in the `cdf modules add` command as that file is required for
that command. Now, we instead create a default `cdf modules add`.

## Changelog

- [x] Patch
- [ ] Minor
- [ ] Skip

## cdf

### Fixed

- You can now run `cdf modules add` without having a pre-existing
`config.[env].yaml` file. The command will now add a default for you,
before adding new modules. This is because all modules have variables
that must be added to a `config.[env].yaml` file.

## templates

No changes.