Release 0.5.0 (#1519)
# Description

Release `0.5.0`. This is done by removing the alpha flags for the 4
features that are planned for v5.

## Changelog

- [ ] Patch
- [x] Minor
- [ ] Skip

## cdf

### Changed

- The `cdf dump datamodel` no longer includes data models that are
global (in system space). You can get the original behavior by using the
`--include-global` flag.
- Toolkit now stores a hash of the credentials of
Workflow/Transformation/Function in the resources such that the resource
is updated when the credentials change.
- For Workflow/Transformations/Function Toolkit no longer falls back to
Toolkit credentials when validation-type != 'dev' in the
`config.[env].yaml`.

### Added
- Toolkit now support for `cdf dump workflow/transformation/group/node`.

## templates

No changes.