[CDF-24461] 🧁 Support Infield Configuration Part 2 (#1532)
# Description

Part 2 of #1531. 

Today Toolkit implicitly supports Infield configurations through
supporting Nodes. This is causing complexity on the user side as the
Infield configuration contains for example dataSetId which the user is
then responsible for looking up across CDF projects. This PR adds
explicit support for the Infield Configuration.

This PR uses the data class from the last PR and creates an Infield
Loader.

## Changelog

- [x] Patch
- [ ] Minor
- [ ] Skip

## cdf

### Added

- (alpha) Support for Infield resource.

## templates

No changes.