[CDF-24561] 👷 Introduce Toolkit Resource Classes (#1540)
# Description

This PR introduces Toolkit resource classes. The difference between
these and the write/read (request/response) class in the cognit-sdk, it
that these are based on pydantic and matches the YAML file format that
toolkit expects. This is similar to the write (request) format in the
PySDK with the following two differences:

* Pydantic based enables giving much better error messages to the end
user when they are not following the spec. See #1537 for a prototype.
* Toolkit does small adjustments to write (request) format, for example,
it uses `dataSetExternalId` instead of `dataSetId`, i.e., using external
IDs instead of internal IDs. Another example is credentials, Toolkit
allows the uses to specify `clientId+clientSecret` and Toolkit will
create the `nonced.


I am only adding the TimeSeries here to instroduce a simple example, I
also do not use it yet in the build command. That will be a separate PR.

## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip