[prep CDF-24745, CDF-24554, CDF-24555] 👷Prepare dump refactoring, Interactive Select (#1591)
# Description

**Context**: Toolkit supports dumping assets/timeseries to csv/parquet
(and yaml). Today this command has a lot of functionality that is needed
in other commands. In this part, I recreate the interactive select logic
used to select dataset/asset hierarchy to select which assets/timeseries
to dump. I will reuse this logic in for example the profile command to
do the same type of selection.

## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip