fix configuration value for workflows, fix bug with datapoints extractor (#1580)
Note to risk reviewer: the code in this PR is not part of the Toolkit
CLI per se, but rather a sample used by the Academy team when they run
Bootcamps. Hence the requirements for testing and typing etc are a bit
more relaxed.

# Description

Fixed Bootcamp workflows using incorrect credentials, fixed bug
introduced in last change for datapoints extractor.

## Changelog

- [x] Patch
- [ ] Minor
- [ ] Skip

## cdf

No changes

## templates

### Fixed

- Bootcamp config.test.yaml had incorrect credentials for workflow
variables
- fixed bug that would miss list of datapoints when list was < 50