[prep CDF-24745, CDF-24554, CDF-24555] 👷Prepare dump refactoring, TableWriter class. (#1589)
# Description

**Context**: Toolkit supports dumping assets/timeseries to csv/parquet
(and yaml). That command will be refactored to support non-asset centric
resources such as Canvas as part of migration support. This PR
introduces a TableWriter class with subclasses for writing to
parquet/csv/yaml which will be used in the refactoring. It is pulled out
in this PR to keep PR size small.

## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip