[CDF-24918] ⏱ Toolkit client set-pending-instance-id (#1613)
# Description

In the upgrade from `TimeSeries` to `CogniteTimeSeries`, we need to link
the newly created `CogniteTimeSeries` nodes to existing asset-centric
timeseries to avoid reingesting all datapoints. This is the first PR in
building this tooling that can do that linking. The tooling will depend
on an alpha feature in the `timeseries` endpoint which accepts setting a
property `pendingInstanceId` that is used as follows.

1. Set `pendindInstanceId` on existing asset-centric timeseries
2. Create the `CogniteTimeSeries` with the node ID set in step 1.
3. The timeseries syncer will then link the newly created
`CogniteTimeSeries` node to the existing asset-centri timeseries from
step 1, instead of creating a new one.

This PR adds support for this endpoint into the internal `ToolkitClient`
which is not exposed to the end-user. This will be used as a basis for a
future `cdf migrate timeseries` command.

## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip