[CDF-24935] 🚅 Robustify and speed up instance writing. (#1627)
# Description

This is a reimplementation of the
`client.data_modeling.instances.apply(...)` method from the cognite-sdk
overcoming the following issues
1. The `cognite-sdk` method is limited to 1 worker, even though the API
support 4. This will be used to run migration which can be millions of
nodes, then, we need the full power of the what the API supports.
2. The retry logic in simply backs off and tries again. Learning from
the upgrade pilots turns out that when writing a lot of nodes, this
often leads to exceptions. Instead, the new method backs off, splits the
load into two and tries again with two requests.
3. Standardized the signature of the function to simply take instances
instead of distinguishing between nodes and edges.


Test on writing 1 million nodes


![image](https://github.com/user-attachments/assets/ef9f372f-effe-48ec-b571-32914f21b33c)


## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip