[CDF-24773] 🪲Build only identifiers (#1637)
# Description

The issue solved in this PR is that the `cdf modules pull` command uses
the `cdf build` command internally, but fails if a resource is missing a
required field. Instead of solving these two bugs individually, I
introduce a new example project with all resources with only identifiers
and build that project. This should only raise warnings and not errors.

Note this PR does not fix the issue, but prepares for the fix.

## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip