[prep CDF-24984] 👷Refactor deploy command Part 1 (#1644)
# Description

**Context** I want to reuse the logic from the deploy command in the
upcoming `cdf migrate prepare` command to deploy a migration model. This
is currently not possible as the deploy command expects resources to be
in a file and not already in memory. This PR is part 1 of refactoring of
the deploy command, such that it becomes easier to read and reuse.

The refactoring I have done in this PR is to split the `.execute()`
method into smaller function for increased readability.
The current version is 106 lines long, while the new one introduce in
this PR is 14.

Note there is zero change to logic, thus not touching any tests.

## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip