[prep CDF-24921] 👌 Raw profile endpoint (#1620)
# Description

**Reviewer**: The reason for the size of this PR is the dataclasses
needed to parse the response from the `profiler/raw` endpoint along with
the tests for these classes. The endpoint implementation is +40 lines
(with+57 for test) and dataclasses +370 (with +205 lines of tests).

**Context**: This is part of developing a `cdf profile raw` command. The
motivation for this command is to support migration from assent-centric
to data modeling by analyzing how data is flowing from raw into the
different cdf resources such as assets, events, and files.

This PR introduced the `profiler/raw` endpoint into the internal client
used by Toolkit. **Note this is not exposed to the end user**. This is a
hidden endpoint that profiles raw tables up to 1 million rows. This is
useful for the new `profile raw` command to list number of rows and
columns and thus give an impression of the amount of, for example,
assets that are coming from a specific raw table. The limitation of 1
million rows as well as timeout of profiling, I plan to handle by just
informing the user that a table have either x number of rows, or more
than 1 million. For analyzing assets I expects this limitation will not
be relevant for 95% of the cases.

## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip