[prep CDF-24984] 👷Refactor deploy command part2 (#1645)
# Description

Stacked on #1644 

**Context** I want to reuse the logic from the deploy command in the
upcoming `cdf migrate prepare` command to deploy a migration model. This
is currently not possible as the deploy command expects resources to be
in a file and not already in memory. This PR is part 1 of refactoring of
the deploy command, such that it becomes easier to read and reuse.

In this PR I have:
1. Split the now renamed method `prepare_resources` into three smaller
methods in the `Resource Worker`.
2. Split the now named `deploy_resource_type` method into smaller
methods. This is increased readability and ease of reuse.
3. Renaming of all the deploy methods in the DeployCommand to reflect
the abstraction level.

## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip