[CDF-25127] 🐛Relative path in cdf modules pull (#1680)
# Description

**Context** In Toolkit, we have a command `cdf modules pull` that allows
you to 'pull' configuration from CDF down to local configurations. For
example, I have a transformations which I have changed in the fusion UI,
and I want to update the SQL from the transformation. I can do this with
the command `cdf modules pull
modules/my_module/transformations/my_local_config.Transformation.yaml`.

Giving a relative path would skip the selection of the resources and it
would not be pulled.

## Changelog

- [x] Patch
- [ ] Minor
- [ ] Skip

## cdf

### Fixed

- The `cdf module pull <path>` no longer skips resources when the given
path is relative.

## templates

No changes.