[prep CDF-24982] 🚅 Implemented fast delete (#1684)
# Description

This is a reimplementation of the
`client.data_modeling.instances.delete(...)` method from the cognite-sdk
overcoming the following issues

1. The `cognite-sdk` method is limited to 1 worker, even though the API
support 2. This will be used in the purge command that potentially
deletes million of nodes, which is likely to happen during the
migration.
2. Standardized the signature of the function to simply take instance
IDs instead of distinguishing between nodes and edges. I need this in
the Canvas wrapper, and expect this to be useful for other cases as
well.


## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip