[prep CDF-24982] 🦿 Canvas industrial dataclasses (#1688)
# Description

Stacked on  <del>#1675 #1676</del>

This is part of a multiple PRs for adding support migration of Canvas
from asset-centric to data modeling.

This PR adds the dataclasses `IndustrialCanvas` in read and write format
with some helper methods. These data classes are container classes for
an entire data class. The helper methods will make it simpler to work
with them.

After this PR there will another PR with a CRUD implementation for these
classes. These dataclasses with the CRUD implementation will lead to a
good abstraction level in the `cdf migrate canvas` implementation.

## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip