[CDF-25148] Migration model round 2 (#1706)
# Description

This updates the migration command to support migration of consumption
apps (for example Canvas) and ingestion.

I am not bumping the version of the model from `v1->v2`. Two reasons for
this. This is an alpha feature, so we can change without notice. In
addition, I checked mixpanel and no-one has deployed the current version
of this model (besides myself) so it will not impact anyone.



![image](https://github.com/user-attachments/assets/00bce2ef-6c8b-489f-8fe4-2d6d66f50780)


## Changelog

- [x] Patch
- [ ] Minor
- [ ] Skip

## cdf

### Changed

- [alpha] The model deployed with `cdf migrate prepare` has now changed
to prepare for support of ingestion and consumption.

## templates

No changes.