[prep CDF-24982] 🫅 Canvas industrial CRUD API (#1689)
# Description

Stacked on  <del>#1675 #1676 #1688 #1684</del>. 

This is part of a multiple PRs for adding support migration of Canvas
from asset-centric to data modeling.

This implements CRUD operations on top of the instance REST endpoints
for Canvases. This enables me to create/retrieve/update/delete canvases
as a single object in the migration, thus making the migration command
must nicer to implement


![image](https://github.com/user-attachments/assets/5032b194-0aa8-41ab-a02d-ee73dbdd938b)


## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip