[prep CDF-25149] 🔨 Introduce scope logic (#1723)
# Description

**Context**: Toolkit has functionality such as profiling and auth that
depends on the access level of the user. This PR prepares for a
`get_scope()` method that returns the scope given an action.

For example, in `cdf profile assets` we can input
`get_scope(AssetsAcl.Action.Read)` and know whether the user has access
to all assets, specific datasets or nothing.

## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip