[CDF-25149]  Get Scope (#1703)
# Description

This will be used in multiple commands to inform the user the scope of
the command they are running. For example,
when profiling an asset hierarchy `cdf profile asset --hierarchy
my_hierarchy` with a token that only has access to `my_dataset`. This
will output a warning, something like:

```python
>>> The profiling will only show assets in the 'my_dataset'. 
```

Example usage: I want to know what spaces I am allowed to read from in a
particular project:


![image](https://github.com/user-attachments/assets/948233ae-ccd8-468c-a0df-98ea30df1136)



## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip