[CDF-25200]🧹Cleanup from future import __annotations__ (#1736)
# Description

We have switched our code base a while back from having lower bound on
`3.8` to now having lower bound on `3.10`. This makes `__from__ future
import annotations` no longer necessary. This PR removes it. A benefit
of this is that AI tools gets (hopefully) less confused in the code base
as it no longer sees 3.9 and earlier syntax for Python.

## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip